Add style changes and type casting Prior to this commit a couple files did not include a newline at the end of the file. In addition to this the uint_8 override field of the incoming messages was being transfered to an int type without proper casting. This commit fixes those issues.

+++ control/joystick_commander/src/can_read.c
@@ -62,7 +62,7 @@ int handle_can_rx_brake_report(
     {
         if( (*driver_override_flag) == 0 )
         {
-            (*driver_override_flag) = brake_report->override;
+            (*driver_override_flag) = (int) brake_report->override;
         }
     }
     
@@ -86,7 +86,7 @@ int handle_can_rx_throttle_report(
     {
         if( (*driver_override_flag) == 0 )
         {
-            (*driver_override_flag) = throttle_report->override; 
+            (*driver_override_flag) = (int) throttle_report->override; 
         }
     }
     
@@ -110,7 +110,7 @@ int handle_can_rx_steering_report(
     {
         if( (*driver_override_flag) == 0 )
         {
-            (*driver_override_flag) = steering_report->override;
+            (*driver_override_flag) = (int) steering_report->override;
         }
     }
